refactor(insights): components
